Updated Dealer & Functionality
I updated the Dealer to play for itself, and cleaned up some code. i.e.
shortened the bust() and win() functions, that now solely rely on the
countHand() function to count everyone's hands. Also, laid some
groundwork for an update that will allow any Player to choose to double
down, if they happen to get a pair on the deal.